# Speed up CH for foot