--- a/core/src/main/java/com/graphhopper/routing/ch/PrepareContractionHierarchies.java
+++ b/core/src/main/java/com/graphhopper/routing/ch/PrepareContractionHierarchies.java
@@ -533,7 +533,7 @@ public class PrepareContractionHierarchies {
                 // todo: optimize
                 return new Params(0, 100, 0, -1, 100, 5);
             } else {
-                return new Params(0, 100, 100, 5, 100, 20);
+                return new Params(0, 100, 100, 2, 100, 20);
             }
         }